+++ __brick__/{{app_name}}/test/widget_test.dart
@@ -8,7 +8,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:macos_ui/macos_ui.dart';
 
-import 'package:{{ app_name }}/main.dart';
+import 'package:{{ app_name.snakeCase() }}/main.dart';
 
 void main() {
   testWidgets('App is built with a MacosWindow parent widget',